backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pillow==10.2.0
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.3
alembic==1.13.1
python-dotenv==1.0.0
